Extract the ExpressionTree Visitor into it's own class. Clean up the demo problems.
